+++ library.properties
@@ -1,10 +1,12 @@
 name=avdweb_VirtualDelay
 version=1.0.3
-author=Albert van Dalen
-maintainer=Albert van Dalen <a@maxun.cc>
+author=Albert van Dalen <http://www.avdweb.nl>
+license=GPL-3.0-or-later
+maintainer=Abhijit Bose (aka. Boseji) <boseji.com>
 sentence=Allows using (multiple) delays without blocking code execution. Arduino Uno and Zero.
 paragraph=During the delay, the code execution is continued. We can use multiple delays simultaneously and independent of each other.
 category=Timing
 url=https://github.com/avandalen/VirtualDelay
 architectures=avr, samd
+includes=avdweb_VirtualDelay.h
